+++ coursehelperapp/src/main/java/com/coursehelper/controllers/HomePageController.java
@@ -1,4 +1,4 @@
-package com.coursehelper;
+package com.coursehelper.controllers;
 
 import javafx.fxml.FXML;
 import javafx.scene.text.Text;
